+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [ ] File globbing pattern support (#49)
 - [ ] Show yellow InfoBar when initializing
 - [ ] Default settings for source maps (#120)
+- [x] Use new VS2015 Error List API
 - [x] MSBuild: Changed build ordering (#124)
 - [x] Compile only when files have changed (#121)
 - [x] Fixed loss of compile on save (#125)